Moved to ManageHearingsWrapper to reduce coupling
